Current status: 	- The leakage-response experiments are backed out of the active path: the signed parametric wall penetration change and rasterized marker generation should not be treated as the next model. 	- Boundary markers are being re-centered conceptually as broad-phase sentinels: they identify cells near a wall, while the wall evaluator owns physical position, normal, penetration, and force response. 	- The next wall model should be function-wall based for the active 2D work: y=f(x) segments with configured outward normals, plus a vertical-wall special case. 	- Vulkan should stay unchanged until the Python function-wall behavior is validated. Code Changes: 	- Documented the function-wall boundary direction in docs/Dynamics/ProblemEvolution.tex. 	- Captured the required implementation order: cleanup first, Python function-wall changes second, Python validation third, Vulkan/export changes last. 	- Preserved the current simple/reservoir checkpoint notes, including dead-particle vertex filtering and ParticleNumber error diagnostics. Next: 	- Clean up the active wall configuration/evaluator shape. 	- Implement and validate function-wall response in Python, especially small particles at the converging transition. 	- After Python passes, update the simple GLSL exporter/templates and rerun Vulkan captures.

+++ vulkan/src_app/ParticleBoundary/CommandParticleBoundaryAndSpheres.cpp
@@ -236,7 +236,7 @@ void CommandParticleBoundaryAndSpheres::RecordSubPassCube(uint32_t imageindex, u
 		upcosize2,
 		sfl2);
 	ResourceVertexParticle* rvp = (ResourceVertexParticle*)(m_RCO->GetResourceName("VertexParticle"));
-	uint16_t vnum2 = dvo2->m_NumElements;
+	uint32_t vnum2 = dvo2->m_NumElements;
 	uint32_t numpts = rvp->m_NumParticles-1;
 	vkCmdDraw(m_CommandBuffers[currentBuffer], vnum2, numpts, 0, 0);
 
@@ -314,7 +314,7 @@ void CommandParticleBoundaryAndSpheres::RecordSubPassParticle(uint32_t imageinde
 		sfl);
 
 
-	uint16_t vnum = dvo->m_NumElements;
+	uint32_t vnum = dvo->m_NumElements;
 	vkCmdDraw(m_CommandBuffers[currentBuffer], vnum, 1, 0, 0);
 	
 }
